+++ kc/matrix.h
@@ -109,8 +109,8 @@ extern uint8_t col_ghost[NumCols], row_ghost[NumRows];  // marked as having ghos
 
 extern int cnt_press, cnt_hold, cnt_rel;  // testing and stats
 extern int16_t cnt_press1min;
-extern uint16_t cnt_press_key[ScanKeys];
-extern uint16_t cnt_press_max;  // highest value in cnt_press_key
+extern uint32_t cnt_press_key[ScanKeys];
+extern uint32_t cnt_press_max;  // highest value in cnt_press_key
 
 
 //  ----- Functions -----
